+++ src/apps/synthetix/optimism/synthetix.perp-v2.contract-position-fetcher.ts
@@ -1,5 +1,5 @@
-
 import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
+
 import { OptimismSynthetixPerpContractPositionFetcher } from './synthetix.perp.contract-position-fetcher';
 
 @PositionTemplate()